@@ -2,29 +2,18 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:pub/app/establishment/blocs/establishment_bloc.dart';
-import 'package:pub/app/establishment/view-models/establishment_view_model.dart';
-import 'package:pub/app/participant/view-models/participant_view_model.dart';
 
+import '../../../room/blocs/room_bloc.dart';
 import '../../../shared/configs/app_colors.dart';
 import '../../../shared/configs/app_images.dart';
-import '../../../shared/configs/app_routes.dart';
-
-import '../../../room/infra/models/dto/room_dto.dart';
 import '../../../room/view-models/room_view_model.dart';
 import 'package:google_fonts/google_fonts.dart';
 
-import '../../domain/use-cases/get_establishments.dart';
-
 class EstablishmentPageOneWidget extends StatefulWidget {
-  /* EstablishmentPageOneWidget(
-      this._establishmentViewModel, this.participantViewModel, this.bloc);
+  final RoomBloc bloc;
+  final RoomViewModel roomViewModel;
 
-  final ParticipantViewModel participantViewModel;*/
-  final EstablishmentBloc bloc;
-  final EstablishmentViewModel establishmentViewModel;
-  EstablishmentPageOneWidget(
-      {required this.establishmentViewModel, required this.bloc});
+  EstablishmentPageOneWidget({required this.roomViewModel, required this.bloc});
   @override
   State<EstablishmentPageOneWidget> createState() =>
       _EstablishmentPageOneWidgetState();
@@ -32,12 +21,9 @@ class EstablishmentPageOneWidget extends StatefulWidget {
 
 class _EstablishmentPageOneWidgetState
     extends State<EstablishmentPageOneWidget> {
-  late final RoomViewModel _roomViewModel;
-  late StreamSubscription mSub;
+  //late StreamSubscription mSub;
   @override
   initState() {
-    _roomViewModel = RoomViewModel();
-
     // final IGetEstablishments _getEstablishments = GetEstablishments();
     //bloc = EstablishmentBloc(getEstablishmentsUseCase: null);
     /*mSub = bloc.stream.listen((state) {
@@ -47,10 +33,8 @@ class _EstablishmentPageOneWidgetState
   }
 
   @override
-  didChangeDependencies() {
-    /* mSub = bloc.stream.listen((state) {
-      if (state is LeavePublicRoomMessageState) this.mSub.cancel();
-    });*/
+  void didChangeDependencies() {
+    // TODO: implement didChangeDependencies
     super.didChangeDependencies();
   }
 
@@ -69,11 +53,11 @@ class _EstablishmentPageOneWidgetState
               borderRadius: BorderRadius.only(
                   topLeft: const Radius.circular(10.0),
                   topRight: const Radius.circular(10.0))),
-          child: BlocBuilder<EstablishmentBloc, EstablishmentState>(
+          child: BlocBuilder<RoomBloc, RoomState>(
               bloc: this.widget.bloc,
               buildWhen: (context, current) =>
                   context.runtimeType != current.runtimeType &&
-                  (current is SuccessEstablishmentsState),
+                  (current is SuccessRoomsState),
               builder: (context, state) {
                 if (state is InitialState) {
                   return Stack(
@@ -85,36 +69,30 @@ class _EstablishmentPageOneWidgetState
                               padding: EdgeInsets.only(top: 270),
                               child: Center(
                                   child: CircularProgressIndicator(
-                                      valueColor: AlwaysStoppedAnimation<Color>(
-                                          AppColors.darkBrown))))
+                                      valueColor: AlwaysStoppedAnimation<Color>(AppColors.darkBrown))))
                         ])
                       ]);
                 }
                 /*else if (state is LoadingRoomsState ) {
                   return Container(child:Center(child:CircularProgressIndicator()));
                 }*/
-                else if (state is SuccessEstablishmentsState) {
+                else if (state is SuccessRoomsState) {
+                  this.widget.roomViewModel.setRooms(this.widget.roomViewModel.calculateDistance(state.entities));
 
-                  this.widget.establishmentViewModel.setEstablishments(state.entities);
-                  
                   return RefreshIndicator(
                       color: AppColors.darkBrown,
                       onRefresh: () async {
-                        this.widget.bloc.add(GetEstablishmentsEvent());
+                        this.widget.bloc.add(GetRoomsEvent());
                       },
                       child: ListView.builder(
                           scrollDirection: Axis.vertical,
                           shrinkWrap: true,
-                          itemCount: this
-                              .widget
-                              .establishmentViewModel
-                              .getEstablishments
-                              .length,
+                          itemCount: this.widget.roomViewModel.getRooms.length,
                           itemBuilder: (context, index) {
                             return Padding(
                                 padding: const EdgeInsets.only(bottom: 12.0),
                                 child: AnimatedBuilder(
-                                    animation: this._roomViewModel,
+                                    animation: this.widget.roomViewModel,
                                     builder: (context, child) {
                                       return ListTile(
                                           leading: Padding(
@@ -125,96 +103,66 @@ class _EstablishmentPageOneWidgetState
                                                   height: 40,
                                                   decoration: BoxDecoration(
                                                       color: AppColors.white,
-                                                      borderRadius: BorderRadius.all(
-                                                          const Radius.circular(
-                                                              5.0)),
+                                                      borderRadius:
+                                                          BorderRadius.all(const Radius.circular(5.0)),
                                                       boxShadow: [
                                                         BoxShadow(
-                                                            color: Colors.grey
-                                                                .withOpacity(
-                                                                    0.15),
+                                                            color: Colors.grey.withOpacity(0.15),
                                                             spreadRadius: 1,
                                                             blurRadius: 3,
-                                                            offset:
-                                                                Offset(1, 3)),
+                                                            offset: Offset(1, 3)),
                                                       ]),
-                                                  child: this
-                                                          .widget
-                                                          .establishmentViewModel
-                                                          .isAcceptedLocation(this
-                                                                  .widget
-                                                                  .establishmentViewModel
-                                                                  .getEstablishments[
-                                                              index])
-                                                      ? Image.asset(AppImages.lightLogo,
-                                                          width: 20, height: 20)
-                                                      : Image.asset(AppImages.lightUnauthorizedLogo,
+                                                  child: this.widget.roomViewModel.getRooms[index].isAcceptedLocation
+                                                      ? Image.asset(
+                                                          AppImages.lightLogo,
+                                                          width: 20,
+                                                          height: 20)
+                                                      : Image.asset(
+                                                          AppImages.lightUnauthorizedLogo,
                                                           width: 20,
                                                           height: 20))),
                                           title: Padding(
                                               padding: EdgeInsets.only(bottom: 10),
-                                              child: this.widget.establishmentViewModel.isAcceptedLocation(this.widget.establishmentViewModel.getEstablishments[index])
-                                                  ? Text(this.widget.establishmentViewModel.getEstablishments[index].getRoomName,
+                                              child: this.widget.roomViewModel.getRooms[index].isAcceptedLocation
+                                                  ? Text(this.widget.roomViewModel.getRooms[index].name,
                                                       style: GoogleFonts.inter(
                                                         color: AppColors.brown,
                                                         fontSize: 18,
                                                       ))
-                                                  : Text(this.widget.establishmentViewModel.getEstablishments[index].name,
+                                                  : Text(this.widget.roomViewModel.getRooms[index].name,
                                                       style: GoogleFonts.inter(
                                                         color: Colors.grey,
                                                         fontSize: 18,
                                                       ))),
                                           subtitle: Row(
                                             children: [
-                                              this
-                                                          .widget
-                                                          .establishmentViewModel
-                                                          //.getEstablishments[
-                                                          //    index]
-                                                          .participants
-                                                          .length ==
-                                                      1
+                                              this.widget.roomViewModel.participants.length == 1
                                                   ? Text(
-                                                      '${this.widget.establishmentViewModel.participants.length} pessoa',
+                                                      '${this.widget.roomViewModel.participants.length} pessoa',
                                                       style: GoogleFonts.inter(
                                                           fontSize: 13,
                                                           color:
                                                               Colors.black45))
                                                   : Text(
-                                                      '${this.widget.establishmentViewModel.participants.length} pessoas',
+                                                      '${this.widget.roomViewModel.participants.length} pessoas',
                                                       style: GoogleFonts.inter(
                                                           fontSize: 13,
-                                                          color:
-                                                              Colors.black45)),
-                                              /*Padding(
+                                                          color: Colors.black45)),
+                                              Padding(
                                                   padding:
                                                       EdgeInsets.only(left: 40),
                                                   child: Text(
-                                                      '${(this.widget.establishmentViewModel.getEstablishments[index].getDistance).toStringAsFixed(2)} km de distância',
+                                                      '${(this.widget.roomViewModel.getRooms[index].distance).toStringAsFixed(2)} km de distância',
                                                       style: GoogleFonts.inter(
                                                           fontSize: 13,
-                                                          color:
-                                                              Colors.black45))))*/
+                                                          color: Colors.black45)))
                                             ],
                                           ),
                                           onTap: () {
-                                            if (this
-                                                .widget
-                                                .establishmentViewModel
-                                                .isAcceptedLocation(this
-                                                        .widget
-                                                        .establishmentViewModel
-                                                        .getEstablishments[
-                                                    index])) {
-                                              bool isUserExist = this
-                                                  .widget
-                                                  .establishmentViewModel
-                                                  .verifyNameUser(this
-                                                      .widget
-                                                      .establishmentViewModel
-                                                      .getEstablishments[index]);
+                                            if (this.widget.roomViewModel.getRooms[index].isAcceptedLocation) {
+                                              bool isUserExist = this.widget.roomViewModel.verifyNameUser(this.widget.roomViewModel.getRooms[index]);
                                               if (!isUserExist) {
-                                                // this._establishmentViewModel.setRoom(this._establishmentViewModel.getRooms[index]);
+                                                this.widget.roomViewModel.setRoom(this.widget.roomViewModel.getRooms[index]);
                                                 /*Navigator.pushNamed(context,
                                                     AppRoutes.publicRoomRoute,
                                                     arguments: RoomDTO(
@@ -224,8 +172,7 @@ class _EstablishmentPageOneWidgetState
                                               } else {
                                                 ScaffoldMessenger.of(context)
                                                     .showSnackBar(const SnackBar(
-                                                        content: Text(
-                                                            'Seu nickname já existe na sala, altere-o para entrar')));
+                                                        content: Text('Seu nickname já existe na sala, altere-o para entrar')));
                                               }
                                             }
                                           });
