@@ -1,11 +1,9 @@
-
 import '../../../../user/infra/models/user_model.dart';
+
 class EstablishmentDTO {
   late UserModel _user;
 
   EstablishmentDTO(this._user);
 
   get getUser => _user;
-
 }
-
