@@ -1,73 +1,68 @@
 part of '../blocs/room_bloc.dart';
 
-abstract class RoomState <T> {
-  RoomState(this.message, this.roomViewModel);
-  final T message;
-  final T roomViewModel;
-}
+abstract class RoomState {}
 class InitialState extends RoomState{
-  InitialState() : super(null, null);
+  InitialState();
 }
 class InitialRoomState extends RoomState{
-  InitialRoomState() : super(null, null);
+  InitialRoomState();
 }
-
-class DontBuildState extends RoomState{
-  DontBuildState() : super(null, null);
+class SuccessRoomsState extends RoomState {
+  final List<RoomModel> entities;
+  SuccessRoomsState(this.entities);
 }
-
 class EnterPublicRoomMessageState extends  RoomState{
-  EnterPublicRoomMessageState({required PublicRoomData message, required RoomViewModel roomViewModel}) : super(message, roomViewModel){
+  EnterPublicRoomMessageState({required PublicRoomData message, required RoomViewModel roomViewModel}){
     roomViewModel.addParticipants(message);
     if(message.getUser.getNickname != roomViewModel.getUser.getNickname)
       roomViewModel.addMessages(message);
   }
 }
 class LeavePublicRoomMessageState extends RoomState{
-  LeavePublicRoomMessageState({required PublicRoomData message, required RoomViewModel roomViewModel}) : super(message, roomViewModel){
+  LeavePublicRoomMessageState({required PublicRoomData message, required RoomViewModel roomViewModel}){
     roomViewModel.removeParticipants(message);
     if(message.getUser.getNickname != roomViewModel.getUser.getNickname)
       roomViewModel.addMessages(message);
   }
 }
 class ReceivePublicMessageState extends RoomState {
-  ReceivePublicMessageState({required MessageData message, required RoomViewModel roomViewModel}) : super(message, roomViewModel){
+  ReceivePublicMessageState({required MessageData message, required RoomViewModel roomViewModel}){
     roomViewModel.addMessages(message);
   }
 }
 class ReceivePrivateMessageState extends RoomState {
-  ReceivePrivateMessageState({required MessageData message, required ParticipantViewModel participantViewModel}) : super(message, participantViewModel){
+  ReceivePrivateMessageState({required MessageData message, required ParticipantViewModel participantViewModel}){
     participantViewModel.addMessages(message);
   }
 }
 
 class TypingMessageState extends RoomState{
-  TypingMessageState() : super(null, null);
+  TypingMessageState();
 }
 
 class StoppedTypingMessageState extends RoomState{
-  StoppedTypingMessageState() : super(null, null);
+  StoppedTypingMessageState();
 }
 
 class SendMessageState extends RoomState{
-  SendMessageState() : super(null, null);
+  SendMessageState();
 }
 class SendPrivateMessageState extends RoomState{
-  SendPrivateMessageState() : super(null, null);
+  SendPrivateMessageState();
 }
 
 class SendingMessageState extends RoomState{
-  SendingMessageState() : super(null, null);
+  SendingMessageState();
 }
 class DisconnectState extends RoomState{
-  DisconnectState() : super(null,null);
+  DisconnectState();
 }
 
 class EditMessageState extends RoomState{
-  EditMessageState() : super(null, null);
+  EditMessageState();
 }
 
 class DeleteMessageState extends RoomState{
-  DeleteMessageState() : super(null, null);
+  DeleteMessageState();
 }
 
