@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:pub/app/room/view-models/room_view_model.dart';
+import '../../../room/view-models/room_view_model.dart';
 import 'package:google_fonts/google_fonts.dart';
 import '../../../shared/configs/app_colors.dart';
 import '../../../room/blocs/room_bloc.dart';
