@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:pub/app/app_widget.dart';
+import './app/app_widget.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:pub/app/Message_bloc_observer.dart';
+import './app/Message_bloc_observer.dart';
 import 'package:firebase_core/firebase_core.dart';
 
 import 'firebase_options.dart';
@@ -9,8 +9,6 @@ import 'firebase_options.dart';
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
   Firebase.initializeApp(options: DefaultFirebaseOptions.currentPlatform);
-  BlocOverrides.runZoned(
-    () => runApp(const AppWidget()),
-    blocObserver: MessageBlocObserver(),
-  );
+  Bloc.observer = MessageBlocObserver();
+  runApp(const AppWidget());
 }
